# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CloudSentinel API",
    "description": "\nRead-only AWS security scanning: discover resources, evaluate security rules, score risk.\n\n**Authentication.** `POST /api/auth/login` returns a short-lived access token; send it as\n`Authorization: Bearer <token>`. A refresh token is set as an httpOnly cookie and exchanged at\n`POST /api/auth/refresh`. Roles: VIEWER (read), ANALYST (+ scans and triage), ADMIN (+ users,\nAWS accounts, audit log). The role a route needs is in its description and in docs/api.md.\n\n**Lists** return one page (`limit`, `offset`); the `X-Total-Count` header holds the total.\n\n**Errors** have a `detail` field. Validation errors (422) never echo the submitted values.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/findings": {
      "get": {
        "tags": [
          "findings"
        ],
        "summary": "List All Findings",
        "description": "Findings, highest risk first by default. Repeat status / severity / category to match\nany of several values, e.g. ?severity=HIGH&severity=CRITICAL. The X-Total-Count header\nholds the number of matching findings (for pagination).",
        "operationId": "list_all_findings_api_findings_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "aws_account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aws Account Id"
            }
          },
          {
            "name": "resource_uuid",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource Uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FindingStatus"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Severity"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "resource_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ResourceType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource Type"
            }
          },
          {
            "name": "rule_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 32
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rule Id"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 32
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "min_risk",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Risk"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "risk",
                "severity",
                "last_detected",
                "first_detected",
                "rule_id",
                "resource_id"
              ],
              "type": "string",
              "default": "risk",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "desc",
              "title": "Order"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FindingSummary"
                  },
                  "title": "Response List All Findings Api Findings Get"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the access token expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but your role does not allow this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/findings/{finding_id}": {
      "get": {
        "tags": [
          "findings"
        ],
        "summary": "Get Finding",
        "description": "One finding with its evidence, the rule's explanation and remediation steps.",
        "operationId": "get_finding_api_findings__finding_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Finding Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingDetail"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the access token expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but your role does not allow this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "findings"
        ],
        "summary": "Change Finding Status",
        "description": "Triage a finding. FALSE_POSITIVE needs a note. A scan that detects a RESOLVED finding\nagain reopens it; ACKNOWLEDGED and FALSE_POSITIVE are kept.",
        "operationId": "change_finding_status_api_findings__finding_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Finding Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindingStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingDetail"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the access token expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but your role does not allow this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Category": {
        "type": "string",
        "enum": [
          "NETWORK",
          "DATA_PROTECTION",
          "IDENTITY",
          "LOGGING"
        ],
        "title": "Category"
      },
...
      "ErrorMessage": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorMessage"
      },
      "FindingDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "aws_account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Aws Account Id"
          },
          "resource_uuid": {
            "type": "string",
            "format": "uuid",
            "title": "Resource Uuid"
          },
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "risk_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score"
          },
          "status": {
            "$ref": "#/components/schemas/FindingStatus"
          },
          "resource_type": {
            "type": "string",
            "title": "Resource Type"
          },
          "resource_id": {
            "type": "string",
            "title": "Resource Id"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "first_detected": {
            "type": "string",
            "format": "date-time",
            "title": "First Detected"
          },
          "last_detected": {
            "type": "string",
            "format": "date-time",
            "title": "Last Detected"
          },
          "last_scan_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Scan Id"
          },
          "evidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Evidence"
          },
          "risk_breakdown": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Breakdown"
          },
          "status_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Note"
          },
          "status_updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Updated At"
          },
          "status_updated_by_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Updated By Id"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "first_scan_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Scan Id"
          },
          "resource_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Name"
          },
          "rule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RulePublic"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "aws_account_id",
          "resource_uuid",
          "rule_id",
          "title",
          "category",
          "severity",
          "risk_score",
          "status",
          "resource_type",
          "resource_id",
          "region",
          "first_detected",
          "last_detected",
          "last_scan_id",
          "evidence",
          "risk_breakdown",
          "status_note",
          "status_updated_at",
          "status_updated_by_id",
          "resolved_at",
          "first_scan_id"
        ],
        "title": "FindingDetail"
      },
      "FindingStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "ACKNOWLEDGED",
          "RESOLVED",
          "FALSE_POSITIVE"
        ],
        "title": "FindingStatus"
      },
      "FindingStatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/FindingStatus"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "FindingStatusUpdate"
      },
      "FindingSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "aws_account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Aws Account Id"
          },
          "resource_uuid": {
            "type": "string",
            "format": "uuid",
            "title": "Resource Uuid"
          },
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "risk_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score"
          },
          "status": {
            "$ref": "#/components/schemas/FindingStatus"
          },
          "resource_type": {
            "type": "string",
            "title": "Resource Type"
          },
          "resource_id": {
            "type": "string",
            "title": "Resource Id"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "first_detected": {
            "type": "string",
            "format": "date-time",
            "title": "First Detected"
          },
          "last_detected": {
            "type": "string",
            "format": "date-time",
            "title": "Last Detected"
          },
          "last_scan_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Scan Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "aws_account_id",
          "resource_uuid",
          "rule_id",
          "title",
          "category",
          "severity",
          "risk_score",
          "status",
          "resource_type",
          "resource_id",
          "region",
          "first_detected",
          "last_detected",
          "last_scan_id"
        ],
        "title": "FindingSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReferencePublic": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "title",
          "url"
        ],
        "title": "ReferencePublic"
      },
...
      "ResourceType": {
        "type": "string",
        "enum": [
          "AWS::Account",
          "AWS::EC2::Instance",
          "AWS::EC2::SecurityGroup",
          "AWS::S3::Bucket",
          "AWS::IAM::User",
          "AWS::IAM::Role",
          "AWS::CloudTrail::Trail"
        ],
        "title": "ResourceType"
      },
...
      "RulePublic": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "resource_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Resource Types"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "remediation": {
            "type": "string",
            "title": "Remediation"
          },
          "severity_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity Note"
          },
          "limitations": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limitations"
          },
          "references": {
            "items": {
              "$ref": "#/components/schemas/ReferencePublic"
            },
            "type": "array",
            "title": "References"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "category",
          "severity",
          "resource_types",
          "description",
          "rationale",
          "remediation",
          "severity_note",
          "limitations",
          "references"
        ],
        "title": "RulePublic"
      },
...
      "Severity": {
        "type": "string",
        "enum": [
          "CRITICAL",
          "HIGH",
          "MEDIUM",
          "LOW"
        ],
        "title": "Severity"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}